+++ packages/site-demo/content/product/components/slider/react/without-min-max.tsx
@@ -6,7 +6,7 @@ const App = ({ theme }: any) => {
     const [value, setValue] = useState(20);
     return (
         <>
-            <Slider max={100} min={0} hideMinMaxlabel theme={theme} value={value} onChange={setValue} />
+            <Slider max={100} min={0} hideMinMaxLabel theme={theme} value={value} onChange={setValue} />
         </>
     );
 };